sonarcloud

+++ example/src/Giraffe.cpp
@@ -31,7 +31,7 @@ void Giraffe::Update() {
         dir.x *= -1;
     }
 
-    auto delta = static_cast<float>(Util::Time::GetDeltaTimeMs());
+    auto delta = Util::Time::GetDeltaTimeMs();
     Util::Transform deltaTransform{
         dir * delta, 0.002F * delta,
         glm::vec2(1, 1) * (std::sin(rotation / 2) + 1.0F) * 100.0F};
